backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pdfplumber==0.10.3
python-docx==1.1.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.2
werkzeug==3.0.1
